renaming and cleaning a bit dependency parsing
